replcaed double with BigDeceminal for money related fields

+++ src/main/java/com/khomishchak/cryproportfolio/model/DepositWithdrawalTransaction.java
@@ -1,6 +1,7 @@
 package com.khomishchak.cryproportfolio.model;
 
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.Builder;
@@ -12,7 +13,7 @@ public class DepositWithdrawalTransaction extends Transaction {
     private TransactionType transactionType;
 
     @Builder(builderMethodName = "depositWithdrawalTransactionBuilder")
-    public DepositWithdrawalTransaction(String transactionId, String transactionHash, String ticker, double fee, double amount, Date createdAt, TransactionType transactionType) {
+    public DepositWithdrawalTransaction(String transactionId, String transactionHash, String ticker, BigDecimal fee, BigDecimal amount, Date createdAt, TransactionType transactionType) {
         super(transactionId, transactionHash, ticker, fee, amount, createdAt);
         this.transactionType = transactionType;
     }
